requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
streamlit>=1.40.0
pydantic>=2.0.0
requests>=2.31.0
pypdf>=4.0.0
python-dotenv>=1.0.0
structlog>=24.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
langsmith>=0.1.0
